backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
alembic
python-multipart
